added delete video with assignment

+++ src/components/admin/assignment/AssignmentItem.jsx
@@ -10,6 +10,10 @@ export default function AssignmentItem({ assignment, index }) {
 
   // delete assignment
   const handleDeleteAssignment = () => {
+    const confirmation = window.confirm('Are you sure you want to delete?');
+    if (!confirmation) {
+      return;
+    }
     deleteAssignment(id);
   };
